more

+++ pkg/fe/prepare-run.go
@@ -21,12 +21,53 @@ type CompileOptions struct {
 	UseThisRunName string
 }
 
-func Compile(backend be.Backend, opts CompileOptions) (ir.Linked, error) {
+// TODO move into RestoreOptions
+func valuesFromShrinkwrap(templatePath string, opts compilation.Options) (compilation.Options, error) {
+	shrinkwrappedOptions, err := compilation.RestoreOptions(templatePath)
+	if err != nil {
+		return opts, err
+	} else {
+		if opts.Namespace == "" {
+			opts.Namespace = shrinkwrappedOptions.Namespace
+		}
+		// TODO here... how do we determine that boolean values were unset?
+		if opts.ImagePullSecret == "" {
+			opts.ImagePullSecret = shrinkwrappedOptions.ImagePullSecret
+		}
+
+		// careful: `--set x=3 --set x=4` results in x having
+		// value 4, so we need to place the shrinkwrapped
+		// options first in the list
+		opts.OverrideValues = append(shrinkwrappedOptions.OverrideValues, opts.OverrideValues...)
+		opts.OverrideFileValues = append(shrinkwrappedOptions.OverrideFileValues, opts.OverrideFileValues...)
+
+		if opts.Queue == "" {
+			opts.Queue = shrinkwrappedOptions.Queue
+		}
+		// TODO here... how do we determine that boolean values were unset?
+		if opts.HasGpuSupport == false {
+			opts.HasGpuSupport = shrinkwrappedOptions.HasGpuSupport
+		}
+		if !opts.CreateNamespace {
+			opts.CreateNamespace = shrinkwrappedOptions.CreateNamespace
+		}
+	}
+
+	return opts, nil
+}
+
+func PrepareForRun(backend be.Backend, opts CompileOptions) (ir.Linked, error) {
 	stageOpts := compilation.StageOptions{}
 	stageOpts.Verbose = opts.Verbose
 	compilationName, templatePath, _, err := compilation.Stage(stageOpts)
 	if err != nil {
 		return ir.Linked{}, err
+	}
+
+	if updatedOpts, err := valuesFromShrinkwrap(templatePath, opts.CompilationOptions); err != nil {
+		return ir.Linked{}, err
+	} else {
+		opts.CompilationOptions = updatedOpts
 	}
 
 	namespace := opts.CompilationOptions.Namespace
@@ -53,7 +94,10 @@ func Compile(backend be.Backend, opts CompileOptions) (ir.Linked, error) {
 		return ir.Linked{}, err
 	}
 
-	defer os.RemoveAll(templatePath)
+	if !opts.Verbose {
+		defer os.RemoveAll(templatePath)
+	}
+
 	if yaml, err := template.Template(runname, namespace, templatePath, yamlValues, template.TemplateOptions{OverrideValues: dashdashSetValues, OverrideFileValues: dashdashSetFileValues, Verbose: opts.Verbose}); err != nil {
 		return ir.Linked{}, err
 	} else if hlir, err := parser.Parse(yaml); err != nil {
